amend file name to groups
